pi-scale-client/requirements.txt
pyserial>=3.5
python-socketio[client]>=5.10
python-dotenv>=1.0
evdev>=1.6
